transaction_service/requirements.txt
grpcio
grpcio-tools
sqlalchemy
pymysql
pytest
